requirements.txt
asgiref==3.3.0
certifi==2021.10.8
Django==3.1.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
mod-wsgi==4.7.1
numpy==1.19.1
pip==21.2.4
psycopg2==2.8.5
pydub==0.25.1
pytz==2020.1
setuptools==58.0.4
sqlparse==0.4.1
wheel==0.37.1
